refactoring component responsibilities
