+++ src/shared/testcases.js
@@ -1,0 +1,16 @@
+//placeholder file for the validation tests I need to run
+//
+// - valid numSeg
+// - blank toRemove
+// + disable submit
+// + numberline visible with no white gaps
+//
+//
+//2.
+// - enter 10 for numSeg
+// - enter 5, 6 for toRemove
+// - change numseg to 4
+// + numberline greyed out
+// + error on toRemove
+// + disable submit
+//
